Only install/configure boot loader when using a custom kernel

+++ 031-configure.sh
@@ -161,11 +161,13 @@ ln -s usr/lib/systemd/system/rpc-gssd.service \
     >/dev/null 2>&1;
 
 # Create configuration and install syslinux bootloader
-echo 'Creating syslinux configuration...';
+if [ ! -f /boot/vmlinuz-linux ]; then
+    # Only install/configure boot loader when using a custom kernel
+    echo 'Creating syslinux configuration...';
 
-rm -fr /boot/syslinux >/dev/null 2>&1;
+    rm -fr /boot/syslinux >/dev/null 2>&1;
 
-cat > /boot/syslinux.cfg << EOF
+    cat > /boot/syslinux.cfg << EOF
 default linux
 prompt 0
 timeout 60
@@ -176,9 +178,10 @@ label linux
     append root=/dev/sda1 ro quiet
 EOF
 
-echo 'Installing bootloader...';
-extlinux -i /boot >/dev/null 2>&1;
-dd if=/usr/lib/syslinux/mbr.bin of=/dev/sda bs=440 count=1 >/dev/null 2>&1;
+    echo 'Installing bootloader...';
+    extlinux -i /boot >/dev/null 2>&1;
+    dd if=/usr/lib/syslinux/mbr.bin of=/dev/sda bs=440 count=1 >/dev/null 2>&1;
+fi;
 
 # Create user and disable root login
 echo "Creating user $USER...";
